add some feauture

+++ server/services/AdminService.php
@@ -52,29 +52,37 @@ public static function GetTraineeInfoDays(mysqli $connection, int $traineeId)
 }
 public static function deleteTrainee($connection,$traineeId){
 $isTraineeFound=User::findById($connection,$traineeId);
+
 if($isTraineeFound==null){
     return "the trainee id is wrong";
 }
+ $isde1=Model::deleteWhere($connection, "meals_log", "user_id", $traineeId);
+ if(!$isde1){
+    return "erorr";
+ }
+ $isde2= Model::deleteWhere($connection, "trainees_days_info", "user_id", $traineeId);
+ if(!$isde2){
+    return "erorr";
+ }
 if(User::deleteById($connection,$traineeId))
     return true;
 return false;
 
 }
-public static function deleteTraineeDayInfo($connection,$day,$traineeid){
-$isTraineeFound=User::findById($connection,$traineeId);
-if($isTraineeFound==null){
-    return "the trainee id is wrong";
-}
-$existTraineeDaInfo=TraineeDayInfo::findAllWhere($connection,["user_id"=>$trainee_id,"day"=>$day]);
-if(empty($existTraineeDaInfo)){
-    return "no data to delete";
-}
-if(existTraineeDaInfo->deleteByObject($connection)){
-    return true;
-}
-return false;
-
-}
+ public static function deleteTraineeDayInfo($connection,$day,$traineeId){
+ $isTraineeFound=User::findById($connection,$traineeId);
+ if($isTraineeFound==null){
+     return "the trainee id is wrong";
+ }
+ $existTraineeDaInfo=TraineeDayInfo::findAllWhere($connection,["user_id"=>$traineeId,"day"=>$day]);
+ if(empty($existTraineeDaInfo)){
+     return "no data to delete";
+ 
+ if(existTraineeDaInfo->deleteByObject($connection)){
+     return true;
+ }
+ return false
+ }
